Update ydl.py with QThread and functionality
Some more minor updates, but configured it more appropriately to the needs of the software. A QThread worker class was instantiated to make sure that the rest of the GUI wouldn't freeze up or hang due to the while loop being used for the queue. Also I think there was also some minor bugs however I can't really confirm this is at the moment.